pyproject.toml
[project]
name = "blog-agent"
version = "0.1.0"
description = "Autonomous Blog Generation Agent using LangGraph and FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "langchain>=1.2.6",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "langsmith>=0.6.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "streamlit>=1.30.0",
    "requests>=2.31.0",
]

[project.scripts]
blog-agent = "blog_agent.api.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
